Inline Svn::Info to prevent using it from other places
