+++ bin/update-version.h.rb
@@ -6,8 +6,55 @@ require "tmpdir"
 vcs, repo_path, *rest = ARGV
 case vcs
 when "svn"
-  require_relative "../lib/svn/info"
-  branches = Svn::Info.new(repo_path, rest.first).branches
+  class SvnInfo
+    def initialize(repo_path, rev)
+      @repo_path = repo_path
+      @revision = Integer(rev)
+      get_changed
+    end
+
+    def branches
+      [*@added_files, *@updated_files].map { |x| x[/((?:branches\/)?.+?)\//, 1] }.uniq
+    end
+
+    private
+
+    def get_changed
+      @added_files = []
+      @updated_files = []
+      svnlook("changed").each_line do |line|
+        if /^(.).  (.*)$/ =~ line
+          modified_type = $1
+          path = $2
+          case modified_type
+          when "A"
+            add_path(path, @added_files)
+          when "D"
+            # noop
+          else
+            add_path(path, @updated_files)
+          end
+        else
+          raise "unknown format: #{line}"
+        end
+      end
+    end
+
+    def add_path(path, files)
+      unless directory_path?(path)
+        files << path
+      end
+    end
+
+    def svnlook(command, *others)
+      `svnlook #{command} #{@repo_path} -r #{@revision} #{others.join(' ')}`.force_encoding("ASCII-8BIT")
+    end
+
+    def directory_path?(path)
+      path[-1] == ?/
+    end
+  end
+  branches = SvnInfo.new(repo_path, rest.first).branches
 when "git"
   branches = rest.each_slice(3).map do |_oldrev, _newrev, refname|
     IO.popen(["git", "rev-parse", "--symbolic", "--abbrev-ref", refname], &:read).strip
